+++ src/components/layout/Layout.js
@@ -8,7 +8,7 @@ import { MainLayout } from "author-design-system-react";
 import { ConfigContext } from "@/context/context";
 
 import { getLayoutProps } from "./layoutSetup";
-import { initAuthRefresh } from "@/utils/auth/auth";
+import { initAuthRefresh } from "@/utils/auth/authRefresh";
 
 export default function Layout({ appConfig, username, children }) {
